use a single modal to manage auth methods
