+++ lib/lightning_web/live/workflow_live/edit.ex
@@ -267,33 +267,14 @@ defmodule LightningWeb.WorkflowLive.Edit do
         </.form>
 
         <.live_component
-          :if={@selected_trigger && @selected_trigger.webhook_auth_methods == []}
-          module={LightningWeb.WorkflowLive.WebhookAuthMethodFormComponent}
-          id="new_webhook_auth_method_modal"
-          action={:new}
-          trigger={@selected_trigger}
-          return_to={
-            ~p"/projects/#{@project.id}/w/#{@workflow.id}?#{%{s: @selected_trigger.id}}"
-          }
-          webhook_auth_method={
-            %Lightning.Workflows.WebhookAuthMethod{project_id: @project.id}
-          }
-        />
-
-        <.live_component
-          :if={
-            @selected_trigger && length(@selected_trigger.webhook_auth_methods) > 0
-          }
-          module={LightningWeb.WorkflowLive.TriggerWebhookAuthMethodsComponent}
-          id={"#{@selected_trigger.id}_webhook_auth_methods_modal"}
+          :if={@selected_trigger}
+          module={LightningWeb.WorkflowLive.WebhookAuthMethodModalComponent}
+          id="webhooks_auth_method_modal"
           action={:new}
           trigger={@selected_trigger}
           project={@project}
           return_to={
-            ~p"/projects/#{@project.id}/w/#{@selected_trigger.workflow_id}?#{%{s: @selected_trigger.id}}"
-          }
-          webhook_auth_method={
-            %Lightning.Workflows.WebhookAuthMethod{project_id: @project.id}
+            ~p"/projects/#{@project.id}/w/#{@workflow.id}?#{%{s: @selected_trigger.id}}"
           }
         />
       </div>
